Twitch streams data analysis
In this project, I try to visualize the characteristics of twitch users using a combination of SQLite dataframe to tidying and selecting specific data, and matplotlib.pyplot to visualize it in Jupyter Notebook. Because of limited file size issue, I can't put every file used in this repository. Luckily, edited CSVs file needed to run in jupyter notebook is available.